relocating works (entities and repositories)
