+++ lib/domain_entities/place_entity.dart
@@ -1,120 +1,26 @@
-import 'dart:convert';
-import 'package:dio/dio.dart';
 import 'package:places/data_repositories/place_repository.dart';
-
 import 'package:places/domain_models/place.dart';
-import 'package:places/data_repositories/mocks.dart';
-import 'package:places/ui_my_app/my_app.dart';
-
-import 'package:places/data_other/exceptions.dart';
 
 class PlaceEntity {
-  PlaceEntity({required PlaceRepository placeRepository}) {
-    final baseOptions = BaseOptions(
-      baseUrl: 'https://test-backend-flutter.surfstudio.ru',
-      connectTimeout: 5000,
-      receiveTimeout: 5000,
-      sendTimeout: 5000,
-      // ignore: avoid_redundant_argument_values
-      responseType: ResponseType.json,
-    );
-    dio = Dio(baseOptions);
+  PlaceEntity({required this.placeRepository});
 
-    dio.interceptors.add(
-      InterceptorsWrapper(
-        onRequest: (options, handler) {
-          print(
-            'Запрос: ${options.method} ${options.path} ${options.queryParameters}',
-          );
-          return handler.next(options);
-        },
-        onResponse: (response, handler) {
-          print('Ответ получен ${response.data.toString().substring(0, 100)} ');
-          return handler.next(response);
-        },
-        onError: (DioError e, handler) {
-          NetworkException(
-            queryPath: '',
-            errorName: 'dio error',
-          );
-          return handler.next(e);
-        },
-      ),
-    );
-  }
+  PlaceRepository placeRepository;
 
-  late final Dio dio;
+  //
 
-  List<Place> loadedPlaces = [];
-  bool isRequestDoneWithError = false;
+  List<Place> get loadedPlaces => placeRepository.loadedPlaces;
+  set loadedPlaces(List<Place> value) => placeRepository.loadedPlaces = value;
+
+  bool get isRequestDoneWithError => placeRepository.isRequestDoneWithError;
+  set isRequestDoneWithError(bool value) =>
+      placeRepository.isRequestDoneWithError = value;
 
   Future<void> loadPlaces() async {
-    if (isDebugMockDataInPlaceOfHttp) {
-      await Future<dynamic>.delayed(const Duration(seconds: 3));
-      loadedPlaces = mocks;
-      isRequestDoneWithError = false;
-      return;
-    } else {
-      const String _path = '/place';
-
-      try {
-        final Response response = await dio.get<String>(
-          _path,
-        );
-        print(response.statusCode);
-        if (response.statusCode != 200) {
-          throw NetworkException(
-            queryPath: _path,
-            errorName: '${response.statusCode} ${response.statusMessage}',
-          );
-        }
-        loadedPlaces = parsePlaces(response.data.toString());
-      } on NetworkException catch (e) {
-        isRequestDoneWithError = true;
-        print(e);
-        return;
-      } catch (e) {
-        isRequestDoneWithError = true;
-        print(e);
-        return;
-      }
-      isRequestDoneWithError = false;
-      return;
-    }
-  }
-
-  List<Place> parsePlaces(String rawJson) {
-    final List postListJson = jsonDecode(rawJson) as List;
-    return postListJson
-        //ignore: avoid_annotating_with_dynamic
-        .map((dynamic placeJson) =>
-            Place.fromJson(placeJson as Map<String, dynamic>))
-        .toList();
+    await placeRepository.loadPlaces();
   }
 
   Future<void> addPlace(Place newPlace) async {
-    if (isDebugMockDataInPlaceOfHttp) {
-      await Future<dynamic>.delayed(const Duration(seconds: 3));
-      isRequestDoneWithError = false;
-      return;
-    } else {
-      const String _path = '/place';
-      try {
-        final Response response =
-            await dio.post<String>(_path, data: newPlace.toJson());
-        if (response.statusCode != 200) {
-          throw NetworkException(
-            queryPath: _path,
-            errorName: '${response.statusCode} ${response.statusMessage}',
-          );
-        }
-      } on NetworkException catch (e) {
-        isRequestDoneWithError = true;
-        print(e);
-        return;
-      }
-      isRequestDoneWithError = false;
-      return;
-    }
+    // ignore: unawaited_futures
+    placeRepository.addPlace(newPlace);
   }
 }
